services/flight-simulator/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
kafka-python==2.2.15
pydantic==2.11.7
